+++ src/main/java/model/piece/Byeong.java
@@ -1,7 +1,5 @@
 package model.piece;
 
-import model.position.Position;
-
 import java.util.List;
 
 public class Byeong extends Piece {
