+++ src/main/java/model/piece/Elephant.java
@@ -1,7 +1,5 @@
 package model.piece;
 
-import model.position.Position;
-
 import java.util.ArrayList;
 import java.util.List;
