+++ src/main/java/model/piece/Guard.java
@@ -1,9 +1,5 @@
 package model.piece;
 
-import model.piece.Piece;
-import model.piece.PieceType;
-import model.position.Position;
-
 import java.util.List;
 
 public class Guard extends Piece {
