+++ src/main/java/model/piece/Jol.java
@@ -1,7 +1,5 @@
 package model.piece;
 
-import model.position.Position;
-
 import java.util.List;
 
 public class Jol extends Piece {
