+++ src/main/java/model/piece/Piece.java
@@ -1,7 +1,5 @@
 package model.piece;
 
-import model.position.Position;
-
 import java.util.List;
 import java.util.Objects;
