+++ src/test/java/model/piece/GeneralTest.java
@@ -3,10 +3,6 @@ package model.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import model.piece.General;
-import model.piece.Piece;
-import model.piece.PieceType;
-import model.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
