+++ src/test/java/model/piece/PiecesTest.java
@@ -1,6 +1,5 @@
 package model.piece;
 
-import model.position.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
