+++ src/test/java/model/piece/PositionTest.java
@@ -1,5 +1,6 @@
-package model.position;
+package model.piece;
 
+import model.piece.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
