+++ src/test/java/model/player/PlayerTest.java
@@ -8,7 +8,7 @@ import model.piece.General;
 import model.piece.Guard;
 import model.piece.Horse;
 import model.piece.Pieces;
-import model.position.Position;
+import model.piece.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
